+++ src/app/doctorCategorie/Store/app-state.ts
@@ -3,5 +3,4 @@ export interface AppState{}
 export interface IAppState<T>
 {
     readonly all: IAppState<T>
-
 }
